Refactor the Env to share code with VS
