+++ Continuous.Client.VisualStudio/MainPadCommand.cs
@@ -9,6 +9,7 @@ using System.ComponentModel.Design;
 using System.Globalization;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
+using EnvDTE;
 
 namespace Continuous.Client.VisualStudio
 {
