+++ Continuous.Client.VisualStudio/packages.config
@@ -19,4 +19,5 @@
   <package id="Microsoft.VisualStudio.Utilities" version="14.1.24720" targetFramework="net452" />
   <package id="Microsoft.VisualStudio.Validation" version="14.1.111" targetFramework="net452" />
   <package id="Microsoft.VSSDK.BuildTools" version="14.1.24720" targetFramework="net452" developmentDependency="true" />
+  <package id="Newtonsoft.Json" version="7.0.1" targetFramework="net452" />
 </packages>
